+++ SysDatCMS/UserMenu.cs
@@ -1,6 +1,8 @@
 ﻿using DevExpress.XtraEditors;
 using SysDatCMS.Classes;
+using SysDatCMS.Enums;
 using System;
+using System.Linq;
 using System.Windows.Forms;
 
 namespace SysDatCMS
@@ -12,12 +14,12 @@ namespace SysDatCMS
             InitializeComponent();
         }
 
-        
+
         private void UpdateBtn_Click(object sender, EventArgs e)
         {
             Article selectedArticle = ArticlesTable.GetRow(ArticlesTable.GetSelectedRows()[0]) as Article;
 
-            if (Article.CreateOrModifyArticle(selectedArticle.Id, selectedArticle.Titolo, selectedArticle.Testo, selectedArticle.Status.Id, selectedArticle.Autore.Id) > 0)
+            if (Article.ModifyArticleWithoutImages(selectedArticle.Id, selectedArticle.Titolo, selectedArticle.Testo, selectedArticle.Status.Id, selectedArticle.Autore.Id))
             {
                 XtraMessageBox.Show("Articolo modificato con successo!", "Modifica articolo", MessageBoxButtons.OK, MessageBoxIcon.Information);
             }
@@ -34,8 +36,7 @@ namespace SysDatCMS
             this.Hide();
             ArticleDetails.ShowDialog();
 
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            FillGrid();
             this.Show();
         }
         private void RequestPublishBtn_Click(object sender, EventArgs e)
@@ -53,8 +54,7 @@ namespace SysDatCMS
             this.Hide();
             createArticle.ShowDialog();
             // quando il form appena aperto viene chiuso, si torna qui
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            FillGrid();
             this.Show();
         }
         private void VisualizeArticlesBtn_Click(object sender, EventArgs e)
@@ -63,8 +63,7 @@ namespace SysDatCMS
             this.Hide();
             visualizeArticles.ShowDialog();
             // quando il form appena aperto viene chiuso, si torna qui
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            FillGrid();
             this.Show();
         }
         private void VisualizeUsersBtn_Click(object sender, EventArgs e)
@@ -73,25 +72,27 @@ namespace SysDatCMS
             this.Hide();
             visualizeUsers.ShowDialog();
             // quando il form appena aperto viene chiuso, si torna qui
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            FillGrid();
             this.Show();
         }
         private void DeleteBtn_Click(object sender, EventArgs e)
         {
-            int Id = (int)ArticlesTable.GetFocusedRowCellValue(colId.FieldName);
-
-            if (Article.DeleteArticleById(Id))
+            DialogResult dialogResult = XtraMessageBox.Show("Sei sicuro di voler eliminare l'articolo?", "Eliminazione", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+            if (dialogResult == DialogResult.Yes)
             {
-                XtraMessageBox.Show("Articolo eliminato con successo!", "Eliminazione articolo", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            }
-            else
-            {
-                XtraMessageBox.Show("Articolo non eliminato, errore", "Eliminazione articolo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                int Id = (int)ArticlesTable.GetFocusedRowCellValue(colId.FieldName);
+
+                if (Article.DeleteArticleById(Id))
+                {
+                    XtraMessageBox.Show("Articolo eliminato con successo!", "Eliminazione articolo", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                else
+                {
+                    XtraMessageBox.Show("Articolo non eliminato, errore", "Eliminazione articolo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                }
             }
 
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            FillGrid();
         }
 
 
@@ -112,27 +113,38 @@ namespace SysDatCMS
                     break;
             }
 
-            PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString();
+            // OLD VERSIONE
+            //PublishedUserArticleNumber.Text = Article.GetNumberOfArticlesOfTheCurrentUser().ToString(); //THE BEST METHOD NAME!!!
 
-            ArticleGrid.DataSource = Article.GetCurrentUserArticles();
-            statusLookUpEditBtn.DataSource = Status.GetStatuses();
+            // OLD VERSION
+            //ArticleGrid.DataSource = Article.GetCurrentUserArticles();
+
+            // NEW VERSION
+            FillGrid();
+        }
+
+        private void FillGrid()
+        {
+            var articles = Article.GetArticles(StatusEnum.all); // significa che li voglio tutti indistintamente dallo stato
+
+            if (articles != null && articles.Any())
+            {
+                var userArticles = articles.Where(w => w.Autore.Id.Equals(CMSState.CurrentUserId)).ToList();
+                ArticleGrid.DataSource = userArticles;
+                PublishedUserArticleNumber.Text = userArticles.Count.ToString();
+
+                statusLookUpEditBtn.DataSource = Status.GetStatuses();
+
+                if (CMSState.CurrentUserRoleId == 1)
+                {
+                    ArticlesTable.Columns["Status"].OptionsColumn.AllowEdit = false;
+                }
+            }
         }
         private void UserMenu_FormClosed(object sender, FormClosedEventArgs e)
         {
             CMSState.CurrentUserId = 0;
             this.DialogResult = DialogResult.OK;
-        }
-
-        private void disabledCellEvents1_ProcessingCell(object sender, DevExpress.Utils.Behaviors.Common.ProcessCellEventArgs e)
-        {
-            if (CMSState.CurrentUserRoleId == 1)
-            {
-                e.Disabled = e.FieldName == "Status" ? true : false;
-                //if (e.FieldName == "Status")
-                //{
-                //    e.Disabled = true;
-                //}
-            }
         }
     }
